move json2dl functions to util
